### Summary

The improvements boil down to three things:
1. Successful updates to anki now hash note contents and save the hash in an org property on the note. Subsequent pushes hash the current contents and compare against the saved hash, only pushing to anki when a difference is found.
2. Pushes now defer exporting note fields until it's determined that the note will be sent to anki. Also added a variable to make it easy to turn off exporting by default.
3. Pushes now batch up all api requests and send them as one call to anki at the end of processing. Some actions queue up further api calls upon successful completion, which are also all batched up for a second round of requests to the api, and so on until there are no more queued requests to process.

The end result is somewhere between a 50% and 100% speed up on pushing new cards and a virtually unlimited speed up when pushing a large corpus of existing notes, most of which haven't changed.

### Benchmarks
#### Scenario: 100 new cards, no exporting
Old version
```
anki-editor-push-notes                                        1           7.136857152   7.136857152
anki-editor--push-note                                        100         6.6887327069  0.0668873270
anki-editor--create-note                                      100         6.6881397940  0.0668813979
anki-editor-api-call-result                                   102         5.6642092069  0.0555314628
anki-editor-api-call                                          102         5.6635051920  0.0555245607
anki-editor--fetch                                            102         5.6286292289  0.0551826394
anki-editor--set-note-id                                      100         1.193064913   0.0119306491
anki-editor-note-at-point                                     100         0.1189815169  0.0011898151
```

New version
```
anki-editor-push-notes                                        1           3.495925223   3.495925223
anki-editor-api-dispatch-queue                                1           3.139614536   3.139614536
anki-editor--set-note-id                                      100         1.2845784160  0.0128457841
anki-editor--set-note-hash                                    100         1.1354448910  0.0113544489
anki-editor-api-call                                          3           0.731150873   0.2437169576
anki-editor--fetch                                            3           0.724643433   0.241547811
anki-editor-api-call-result                                   2           0.13582163    0.067910815
anki-editor--draw-progress-bar                                300         0.1276257599  0.0004254191
```
#### Scenario: 99 unchanged cards, 1 changed, no exporting
Old version
```
anki-editor-push-notes                                        1           12.861965705  12.861965705
anki-editor-api-call-result                                   202         12.598017311  0.0623664223
anki-editor-api-call                                          202         12.596747883  0.0623601380
anki-editor--fetch                                            202         12.524509685  0.0620025231
anki-editor--push-note                                        100         12.481670745  0.1248167074
anki-editor--update-note                                      100         12.481030434  0.1248103043
```

New version
```
anki-editor-push-notes                                        1           0.581315192   0.581315192
anki-editor-api-call                                          5           0.321204258   0.0642408516
anki-editor--fetch                                            5           0.319538121   0.0639076242
anki-editor-api-dispatch-queue                                3           0.2021955790  0.0673985263
anki-editor-api-call-result                                   2           0.130345123   0.0651725615
```
#### Scenario: 100 new cards, exporting
Old version
```
anki-editor-push-notes                                        1           15.628968179  15.628968179
anki-editor-note-at-point                                     100         7.9138981689  0.0791389816
anki-editor--export-string                                    200         7.7812412609  0.0389062063
anki-editor--push-note                                        100         7.423356627   0.0742335662
anki-editor--create-note                                      100         7.4226246080  0.0742262460
anki-editor-api-call-result                                   102         6.6357773909  0.0650566410
anki-editor-api-call                                          102         6.6351431740  0.0650504232
anki-editor--fetch                                            102         6.601921938   0.0647247248
anki-editor--set-note-id                                      100         0.9226533030  0.0092265330
```
New version
```
anki-editor-push-notes                                        1           10.63572962   10.63572962
anki-editor--process-note                                     100         7.0228366299  0.0702283662
anki-editor--enqueue-create-note                              100         7.0101047570  0.0701010475
anki-editor-api--note                                         100         7.0059367729  0.0700593677
anki-editor--export-fields                                    100         7.0047806490  0.0700478064
anki-editor--export-string                                    200         6.997352285   0.0349867614
anki-editor-api-dispatch-queue                                1           3.284697947   3.284697947
anki-editor--set-note-id                                      100         1.5486320809  0.0154863208
anki-editor--set-note-hash                                    100         0.9447315479  0.0094473154
anki-editor-api-call                                          3           0.781342395   0.260447465
anki-editor--fetch                                            3           0.774590947   0.2581969823
anki-editor--draw-progress-bar                                300         0.1329500850  0.0004431669
anki-editor-api-call-result                                   2           0.115276907   0.0576384535
```
#### Scenario: 99 unchanged cards, 1 changed, exporting
Old version
```
anki-editor-push-notes                                        1           19.133579662  19.133579662
anki-editor-api-call-result                                   202         11.680356543  0.0578235472
anki-editor-api-call                                          202         11.679067128  0.0578171640
anki-editor--fetch                                            202         11.611660435  0.0574834675
anki-editor--push-note                                        100         11.545812984  0.1154581298
anki-editor--update-note                                      100         11.545062166  0.1154506216
anki-editor-note-at-point                                     100         7.2823560520  0.0728235605
anki-editor--export-string                                    200         6.9578029099  0.0347890145
```
New version
```
anki-editor-push-notes                                        1           0.544985199   0.544985199
anki-editor-api-call                                          5           0.275446766   0.0550893532
anki-editor--fetch                                            5           0.273730727   0.0547461454
anki-editor-api-dispatch-queue                                3           0.165724849   0.0552416163
anki-editor-api-call-result                                   2           0.1212798710  0.0606399355
```
#### Informal Scenario: 999 unchanged notes, 1 changed
I didn't think to instrument this one, unfortunately but the new version took about 2s, whereas, judging by the benchmarks above (and assuming the old version is linear-time in the number of notes...) it would be about 120s in the old version.

#### Methodology
I used elp to get these benchmarks and included every function with more than .1s total runtime. The columns from left to right are: function name, times called, total runtime, runtime per call.

### Future
Exporting is now the major factor by far when sending cards, making up about 70% of the total runtime. Setting the note ID and hash combine for another 25%.

Most of the exporting work seems to be done on the org side so I don't know how realistic it will be to tune that. Maybe it's possible to tweak how we make the calls though.

In any case it's much easier to use anki-editor for maintaining large collections of notes now.