@jonprairie 2024-09-24
It looks like the new hash and marker fields of the note struct caused a mismatch in a few of the tests. I updated the tests to include nil in these fields to match get-note-at-point.