@jonprairie 2024-09-24
First off, I didn't realize I could run these tests locally, so sorry about that. I've got them all passing on my machine now. There were two issues:
1. The test data assumed that the note fields would be exported immediately when reading note-at-point.
2. note-at-point makes a marker at point now, which wasn't matching up with the test data.